Continue graphics implementation
! Camera controls
! Object picking
? Line coordinate problems
? Selection transform

Working on a few things at once. Trying to improve and extend the
interface between the GUI and the actual Qt3DWindow. Currently trying to
implement entity selection and highlighting, etc. Several obvious bugs
with lots of things, should see them dwindle over the days, though.

The camera controller needs a lot of work. The orbit is weird,
translation is weird, zooming is way too slow, etc. Maybe try some sort
of log scaling.

Object picking comes soon -- currently am sort of doing this through the
entity list docked to the GUI, but eventually also want to be able to
click actual body in the 3D window.

When in LINE drawing mode, there is a bizarre coordinate problem. I
mean, it isn't really that bizarre since I know exactly why it is
happening, but it does look strange. For future posterity: improper cast
from screen coordinates to world coordinates.

The current selection highlight mechanism is basically supposed to
replace the selected line with a small-radius (yet bigger than that of
the line) cylinder of some bright color. This works, sort of, and they
can be switched between no problem. The cylinder just appears at the
origin though, which of course I knew would happen because I didn't
apply any transform to it. I just didn't want to work on that right now.
That's next, though.