Implement extrusion -- sort of
Extrusion works, more or less. There is a hyperhack in here for demo
purposes, just so I could see if the rest of the reqs were working,
which they seem to be.

The current issue seems to have something to do with how I'm passing
around the point vector. Probably some sort of cursed pointer issue.